chore: clippy back to silent, and two of the forty-seven were not style
Forty-seven warnings had accumulated. Most were mechanical and `--fix` took
them: collapsible `if let` chains, a needless borrow, items sitting after a
test module. Two were not.

**A `#[test]` written twice** in `scour-gui/src/link.rs`. Harmless as it
stands, and the same shape broke a build in this repository once before, which
is why the lint exists and why it is worth keeping at zero: a warning nobody
reads is a warning that cannot warn.

**A `_ => 0` that no value can reach**, colouring roles in the query field.
Removing it is not tidying. Every role is now named, so a role added to
`scour_core` stops this compiling instead of quietly taking colour zero — the
ordinary ink, and therefore the one mistake nobody would see.

Two redundant `..Default::default()` went the same way and for the same
reason. `ScanRequest` has one field today; a second one appearing should stop
the sweep's call site compiling rather than inherit a value nobody chose.

The twelve test fixtures that assigned `app.room` after `App::default()` are a
helper now — `app(10)` — which says what it is for: `App::default()` has no
room, and a view zero rows tall answers every question about scrolling with the
same number.

The six `large_enum_variant` warnings are left, and told to be quiet with the
reason written beside each. Boxing is wrong in all six: a `Test` is evaluated
per row of a two-million-row walk, and the rest are one per request or one per
key press, where the allocation boxing adds is one more than the message
already made.

710 tests pass, `--all-targets` is silent, `fmt --check` clean.

+++ crates/scour-index-native/tests/whole.rs
@@ -4589,7 +4589,7 @@ fn a_reach_over_segments_that_share_no_dates_still_lands_on_the_page() {
             .map(|i| {
                 entry(
                     &format!("/corpus/{era}/file{i:06}.txt"),
-                    NOW - era * 31_536_000 - i as i64,
+                    NOW - era * 31_536_000 - i,
                     (era * 100_000 + i) as u64 + 1,
                 )
             })
